backend/requirements.txt
# BookIT Backend - Requirements

# Core Django
Django==4.2.7
djangorestframework==3.14.0

# Authentication
djangorestframework-simplejwt==5.3.0

# CORS Support
django-cors-headers==4.3.0

# Filtering
django-filter==23.5

# Image Processing
Pillow==10.1.0

# Environment Variables
python-decouple==3.8

# PostgreSQL Support (for production)
psycopg2-binary==2.9.9

# Date/Time Utilities
python-dateutil==2.8.2

# API Documentation (Optional but recommended)
drf-yasg==1.21.7

# Development Dependencies
# Uncomment these for development
# ipython==8.18.1
# django-debug-toolbar==4.2.0
# django-extensions==3.2.3
